+++ _plugins/data_filters.rb
@@ -148,6 +148,13 @@ module HistoricalDiary
       dates.uniq.sort
     end
 
+    def dates_for_place(key)
+      dates = @context.registers[:site]
+        .data['dates_for_places_keys'][key]
+      return [] if dates.nil?
+      dates.uniq.sort
+    end
+
     def dates_for_source(key)
       dates = @context.registers[:site]
         .data['dates_for_sources_keys'][key]